KEYCLOAK-4523 Refactor core password hash functions
